requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
pytesseract>=0.3.10
googletrans>=3.1.0a0
pyautogui>=0.9.54
pynput>=1.7.6
mss>=9.0.1 